backend/requirements.txt
argon2-cffi==19.1.0
cffi==1.13.0
Django==2.2.8
django-environ==0.4.5
pycparser==2.19
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
psycopg2-binary==2.8.4
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.2
django-cors-headers==3.2.0
drf-writable-nested==0.5.3
django-ratelimit==2.0.0
graphene==2.1.8
graphene-django==2.7.1
django-filter==2.2.0
graphene-django-extras==0.4.8
zappa==0.48.2
zappa-django-utils==0.4.0
graphene-django-cud==0.2.1
